physics/cyclo_earth/requirements.txt
# Cyclo-Earth — Python Dependencies
# ===================================
# pip install -r requirements.txt

fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
paho-mqtt>=2.0.0
numpy>=1.26.0
